// uptimer-client/src/interfaces/monitor.interface.ts
export interface IMonitorDocument {
  id?: number | string;
  monitorId?: number;
  notificationId?: number;
  name: string;
  active?: boolean;
  status?: number;
  userId: number;
  frequency: number;
  alertThreshold: number | string;
  url?: string;
  type: string;
  lastChanged?: Date | string;
  timeout?: number | string;
  uptime?: number;
  redirects?: number | string;
  method?: string;
  headers?: string;
  body?: string;
  httpAuthMethod?: string;
  basicAuthUser?: string;
  basicAuthPass?: string;
  bearerToken?: string;
  contentType?: string;
  statusCode?: string;
  responseTime?: string | number;
  connection?: string;
  port?: number | string;
  heartbeats?: IHeartbeat[];
  notifications?: INotification;
}
// ...
export interface IHeartbeat {
  id?: number;
  monitorId: number;
  status: number;
  code: number;
  message: string;
  timestamp: number | string;
  reqHeaders?: string;
  resHeaders?: string;
  reqBody?: string;
  resBody?: string;
  responseTime: number;
  connection?: string;
}

export interface IPagination {
  start: number;
  end: number;
}
// ...
export interface IMonitorState {
  showModal: boolean;
  enableRefresh: boolean;
  autoRefreshLoading: boolean;
}
// ...
export interface IUseHome {
  monitorState: IMonitorState;
  monitors: IMonitorDocument[];
  limit: IPagination;
  isRefreshed: boolean | null;
  autoMonitorsRef: MutableRefObject<IMonitorDocument[]>;
  monitorsRef: MutableRefObject<IMonitorDocument[]>;
  openModal: boolean | null;
  view: string;
  loading: boolean;
  setView: Dispatch<SetStateAction<"box" | "list">>;
  setMonitors: Dispatch<SetStateAction<IMonitorDocument[]>>;
  updateLimit: (newLimit: IPagination) => void;
  setMonitorState: Dispatch<SetStateAction<IMonitorState>>;
  refreshMonitors: () => Promise<void>;
  enableAutoRefresh: () => Promise<void>;
  closeUptimeModal: () => void;
}